fix(openai): align gpt55 helper metadata
