+++ src/sdk/config.rs
@@ -162,9 +162,9 @@ impl SdkConfigBuilder {
             api_key: api_key.to_string(),
             base_url: None,
             models: vec![
+                "gpt-5.4".to_string(),
                 "gpt-5.5".to_string(),
                 "gpt-5.5-pro".to_string(),
-                "gpt-5.4".to_string(),
                 "gpt-5.4-mini".to_string(),
                 "gpt-5.4-nano".to_string(),
             ],
@@ -588,6 +588,10 @@ mod tests {
                 .models
                 .iter()
                 .any(|model| model.starts_with("gpt-"))
+        );
+        assert_eq!(
+            config.providers[0].models.first(),
+            Some(&"gpt-5.4".to_string())
         );
         assert!(config.providers[0].models.contains(&"gpt-5.4".to_string()));
     }
